+++ ent/schema/human.go
@@ -5,7 +5,7 @@ import (
 	"entgo.io/ent/schema/edge"
 	"entgo.io/ent/schema/index"
 
-	cfmixin "github.com/grokify/coreforge/identity/ent/mixin"
+	cfmixin "github.com/grokify/systemforge/identity/ent/mixin"
 )
 
 // Human holds the schema definition for human-specific principal data.