+++ ent/schema/license.go
@@ -6,7 +6,7 @@ import (
 	"entgo.io/ent/schema/edge"
 	"entgo.io/ent/schema/field"
 	"entgo.io/ent/schema/index"
-	cfmixin "github.com/grokify/coreforge/identity/ent/mixin"
+	cfmixin "github.com/grokify/systemforge/identity/ent/mixin"
 )
 
 // License holds the schema definition for marketplace licenses.